backend/requirements.txt
fastapi[standard]>=0.110.0
motor>=3.3.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
